cleaned up pytests
