+++ MANIFEST.in
@@ -1,14 +1,14 @@
-include chainladder/data/ABC.pkl
-include chainladder/data/auto.pkl
-include chainladder/data/casresearch.pkl
-include chainladder/data/GenIns.pkl
-include chainladder/data/liab.pkl
-include chainladder/data/M3IR5.pkl
-include chainladder/data/mcl.pkl
-include chainladder/data/Mortgage.pkl
-include chainladder/data/MW2008.pkl
-include chainladder/data/MW2014.pkl
-include chainladder/data/quarterly.pkl
-include chainladder/data/RAA.pkl
-include chainladder/data/UKMotor.pkl
-include chainladder/data/USAA.pkl
+include chainladder/utils/data/ABC.pkl
+include chainladder/utils/data/auto.pkl
+include chainladder/utils/data/casresearch.pkl
+include chainladder/utils/data/GenIns.pkl
+include chainladder/utils/data/liab.pkl
+include chainladder/utils/data/M3IR5.pkl
+include chainladder/utils/data/mcl.pkl
+include chainladder/utils/data/Mortgage.pkl
+include chainladder/utils/data/MW2008.pkl
+include chainladder/utils/data/MW2014.pkl
+include chainladder/utils/data/quarterly.pkl
+include chainladder/utils/data/RAA.pkl
+include chainladder/utils/data/UKMotor.pkl
+include chainladder/utils/data/USAA.pkl
